backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
python-frontmatter==1.1.0
markdown==3.7
aiosqlite==0.20.0
python-dotenv==1.0.1
watchfiles==0.24.0
websockets==13.1
PyYAML==6.0.2
bcrypt==4.1.2
PyJWT==2.8.0
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow==10.4.0
cryptography==42.0.0
